requirements.txt
APLpy==2.0.3
astropy==4.3.1
astropy-healpix==0.6
click==8.0.1
cycler==0.10.0
Cython==0.29.24
imageio==2.9.0
importlib-metadata==4.6.4
kiwisolver==1.3.1
matplotlib==3.4.3
networkx==2.6.2
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
progressbar==2.5
PyAVM==0.9.5
pyerfa==2.0.0
pyparsing==2.4.7
pyregion==2.1.1
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
reproject==0.8
scikit-image==0.18.2
scipy==1.7.1
Shapely==1.7.1
six==1.16.0
tifffile==2021.8.8
typing-extensions==3.10.0.0
zipp==3.5.0
